two big changes:
A.had to solve the problem of the edge case: "what if
there are no points from each of the groupings in the
first iteration":
solution: find medoids for each iteration then find medoids
 for medoids of all iterations.
thus, we ensure finding the best medoids without doing the
operation over all the input

that leads to inefficiencies:
augmenting the input twice - once for finding model,
the second for prediction. augmentation takes a long
time if columns have a mix of numbers and other objects.

not sure what to do about it.

B.had to rework how replacement dictionary worked, first
 go column by column, save all unique values.
then if there is a mix of numbers and objects - if there
are more than 80% numbers - dictionary all non-object
numbers to zeros, and numbers to floats (because they are stored as str or objects).
this creates a very big dictionary.